refactor: Move pytest fixtures to its own file
